Start extracting a bunch of logic into separate controls / logical units
