+++ rpkg-gui/App.xaml.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Windows;
+﻿using System.Windows;
 
 namespace rpkg
 {
@@ -13,5 +7,6 @@ namespace rpkg
     /// </summary>
     public partial class App : Application
     {
+        public static UserSettings Settings = new UserSettings();
     }
 }
