+++ rpkg-gui/MainWindow.xaml
@@ -7,7 +7,8 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:wf="clr-namespace:System.Windows.Forms;assembly=System.Windows.Forms"
-        xmlns:local="clr-namespace:rpkg"
+        xmlns:local="clr-namespace:rpkg" 
+        xmlns:controls="clr-namespace:rpkg.Controls"
         mc:Ignorable="d"
         Loaded="Window_Loaded"
         Title="RPKG v2.16" Height="800" Width="1280" WindowStartupLocation="CenterScreen">
@@ -98,21 +99,21 @@
             <mah:ToggleSwitch Grid.Row="0" Grid.Column="4" Toggled="DiscordRichPresence_Toggled" Width="90"/>
         </Grid>
         <TabControl x:Name="LeftTabControl" Grid.Row="1" Grid.Column="0">
-            <TabItem x:Name="FirstTabLeft" Header="Resource View">
+            <TabItem x:Name="TabResourceView" Header="Resource View">
                 <!--TreeView x:Name="MainTreeView" VirtualizingStackPanel.IsVirtualizing="True" VirtualizingStackPanel.VirtualizationMode="Recycling" BorderThickness="1"/-->
                 <WindowsFormsHost x:Name="MainTreeViewWindowsFormsHost">
                     <wf:TreeView x:Name="MainTreeView" Font="Consolas,10"/>
                 </WindowsFormsHost>
             </TabItem>
-            <TabItem x:Name="SecondTabLeft" Header="Dependency View">
+            <TabItem x:Name="TabDependencyView" Header="Dependency View">
                 <!--TreeView x:Name="SearchRPKGsTreeView" VirtualizingStackPanel.IsVirtualizing="True" VirtualizingStackPanel.VirtualizationMode="Recycling" BorderThickness="1"/-->
                 <WindowsFormsHost x:Name="HashMapTreeViewWindowsFormsHost">
                     <wf:TreeView x:Name="HashMapTreeView" Font="Consolas,10"/>
                 </WindowsFormsHost>
             </TabItem>
-            <TabItem x:Name="ThirdTabLeft" Header="Search">
+            <TabItem x:Name="TabSearch" Header="Search">
                 <TabControl x:Name="LeftTabControl2" Grid.Row="1" Grid.Column="0">
-                    <TabItem x:Name="ThirdTabLeft2" Header="Search RPKGs">
+                    <TabItem x:Name="TabSearchRpkg" Header="Search RPKGs">
                         <Grid>
                             <Grid.RowDefinitions>
                                 <RowDefinition Height="Auto"/>
@@ -145,7 +146,7 @@
                             </WindowsFormsHost>
                         </Grid>
                     </TabItem>
-                    <TabItem x:Name="ThirdTabLeft3" Header="Search Hash List">
+                    <TabItem x:Name="TabSearchHashList" Header="Search Hash List">
                         <Grid>
                             <Grid.RowDefinitions>
                                 <RowDefinition Height="Auto"/>
@@ -183,42 +184,19 @@
         </TabControl>
         <GridSplitter HorizontalAlignment="Right" VerticalAlignment="Stretch" Grid.Row="1"  Grid.Column="1" ResizeBehavior="PreviousAndNext" Width="5"/>
         <TabControl x:Name="RightTabControl" Grid.Row="1" Grid.Column="2">
-            <TabItem x:Name="FirstTabRight" Header="Details">
-                <Grid>
-                    <Grid.RowDefinitions>
-                        <RowDefinition Height="Auto" />
-                        <RowDefinition Height="*"/>
-                    </Grid.RowDefinitions>
-                    <Grid.ColumnDefinitions>
-                        <ColumnDefinition Width="Auto"/>
-                        <ColumnDefinition Width="Auto"/>
-                        <ColumnDefinition Width="*"/>
-                    </Grid.ColumnDefinitions>
-                    <Button Grid.Row="0" Grid.Column="0" Click="EditPatchDeletionListButton_Click" x:Name="EditPatchDeletionListButton" Content="Edit Patch Deletion List" Visibility="Collapsed"/>
-                    <Button Grid.Row="0" Grid.Column="0" Click="SaveChangesToPatchDeletionListButton_Click" x:Name="SaveChangesToPatchDeletionListButton" Content="Save Changes To Patch Deletion List" Visibility="Collapsed"/>
-                    <Button Grid.Row="0" Grid.Column="1" Click="DiscardChangesToPatchDeletionListButton_Click" x:Name="DiscardChangesToPatchDeletionListButton" Content="Discard Changes To Patch Deletion List" Visibility="Collapsed"/>
-                    <Button Grid.Row="0" Grid.Column="0" Click="EditHashDependsButton_Click" x:Name="EditHashDependsButton" Content="Edit Hash Depends" Visibility="Collapsed"/>
-                    <Button Grid.Row="0" Grid.Column="0" Click="SaveChangesToHashDependsButton_Click" x:Name="SaveChangesToHashDependsButton" Content="Save Changes To Hash Depends" Visibility="Collapsed"/>
-                    <Button Grid.Row="0" Grid.Column="1" Click="DiscardChangesToHashDependsButton_Click" x:Name="DiscardChangesToHashDependsButton" Content="Discard Changes To Hash Depends" Visibility="Collapsed"/>
-                    <Button Grid.Row="0" Grid.Column="1" Click="ExtractHashDependsButton_Click" x:Name="ExtractHashDependsButton" Content="Extract Hash Depends" Visibility="Collapsed"/>
-                    <TextBox Grid.Row="1" Grid.ColumnSpan="3" x:Name="DetailsTextBox" FontFamily="Consolas" VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Auto" TextWrapping="NoWrap" TextChanged="DetailsTextBox_TextChanged"/>
-                    <ScrollViewer Grid.Row="1" Grid.ColumnSpan="3" x:Name="MainScrollViewer" VerticalScrollBarVisibility="Auto" Visibility="Collapsed">
-                        <StackPanel x:Name="MainStackPanel" Visibility="Collapsed">
-
-                        </StackPanel>
-                    </ScrollViewer>
-                </Grid>
+            <TabItem x:Name="TabDetails" Header="Details">
+                <controls:ItemDetails x:Name="ItemDetails" UnloadRpkg="ItemDetails_OnUnloadRpkg" />
             </TabItem>
-            <TabItem x:Name="SecondTabRight" Header="Hex Viewer">
+            <TabItem x:Name="TabHexViewer" Header="Hex Viewer">
                 <TextBox x:Name="HexViewerTextBox" FontFamily="Consolas" VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Auto" TextWrapping="NoWrap"/>
             </TabItem>
-            <TabItem x:Name="ThirdTabRight" Header="JSON Viewer" Visibility="Collapsed">
+            <TabItem x:Name="TabJsonViewer" Header="JSON Viewer" Visibility="Collapsed">
                 <TextBox x:Name="LocalizationTextBox" FontFamily="Consolas" VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Auto" TextWrapping="Wrap"/>
             </TabItem>
-            <TabItem x:Name="FourthTabRight" Header="Image Viewer" Visibility="Collapsed">
+            <TabItem x:Name="TabImageViewer" Header="Image Viewer" Visibility="Collapsed">
                 <Image x:Name="ImageViewer"/>
             </TabItem>
-            <TabItem x:Name="FifthTabRight" Header="Audio Player" Visibility="Collapsed">
+            <TabItem x:Name="TabAudioPlayer" Header="Audio Player" Visibility="Collapsed">
                 <Grid>
                     <Grid.Resources>
                         <Style TargetType="{x:Type iconPacks:PackIconMaterialDesign}">
@@ -246,7 +224,7 @@
                     </ComboBox>
                 </Grid>
             </TabItem>
-            <TabItem x:Name="SixthTabRight" Header="3D Viewer" Visibility="Collapsed">
+            <TabItem x:Name="Tab3dViewer" Header="3D Viewer" Visibility="Collapsed">
                 <Grid>
                     <hx:HelixViewport3D x:Name="helixViewport">
                         <hx:HelixViewport3D.Camera>
