+++ rpkg-gui/rpkg-gui.csproj
@@ -98,6 +98,9 @@
     <Compile Include="ColorPicker.xaml.cs">
       <DependentUpon>ColorPicker.xaml</DependentUpon>
     </Compile>
+    <Compile Include="Controls\ItemDetails.xaml.cs">
+      <DependentUpon>ItemDetails.xaml</DependentUpon>
+    </Compile>
     <Compile Include="DownloadExtractionProgress.xaml.cs">
       <DependentUpon>DownloadExtractionProgress.xaml</DependentUpon>
     </Compile>
@@ -124,7 +127,13 @@
     </Compile>
     <Compile Include="RpkgLib.cs" />
     <Compile Include="UserSettings.cs" />
+    <Compile Include="Utils\HashUtils.cs" />
+    <Compile Include="Utils\WindowUtils.cs" />
     <Page Include="ColorPicker.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Controls\ItemDetails.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
